log short: Extract rendering logic into a shared type
This will be re-used with the upcoming 'log long' command.

[skip changelog] No user-facing changes.
